+++ llama_index/agent/retriever_openai_agent.py
@@ -1,6 +1,6 @@
 """Retriever OpenAI agent."""
 
-from typing import List, Optional
+from typing import List, Type, Optional
 
 from llama_index.agent.openai_agent import (
     DEFAULT_MAX_FUNCTION_CALLS,
@@ -11,6 +11,7 @@ from llama_index.agent.openai_agent import (
 from llama_index.callbacks.base import CallbackManager
 from llama_index.llms.base import ChatMessage
 from llama_index.llms.openai import OpenAI
+from llama_index.memory import BaseMemory, ChatMemoryBuffer
 from llama_index.objects.base import ObjectRetriever
 from llama_index.tools.types import BaseTool
 
@@ -26,7 +27,7 @@ class FnRetrieverOpenAIAgent(BaseOpenAIAgent):
         self,
         retriever: ObjectRetriever[BaseTool],
         llm: OpenAI,
-        chat_history: List[ChatMessage],
+        memory: BaseMemory,
         prefix_messages: List[ChatMessage],
         verbose: bool = False,
         max_function_calls: int = DEFAULT_MAX_FUNCTION_CALLS,
@@ -34,7 +35,7 @@ class FnRetrieverOpenAIAgent(BaseOpenAIAgent):
     ) -> None:
         super().__init__(
             llm=llm,
-            chat_history=chat_history,
+            memory=memory,
             prefix_messages=prefix_messages,
             verbose=verbose,
             max_function_calls=max_function_calls,
@@ -48,6 +49,8 @@ class FnRetrieverOpenAIAgent(BaseOpenAIAgent):
         retriever: ObjectRetriever[BaseTool],
         llm: Optional[OpenAI] = None,
         chat_history: Optional[List[ChatMessage]] = None,
+        memory: Optional[BaseMemory] = None,
+        memory_cls: Type[BaseMemory] = ChatMemoryBuffer,
         verbose: bool = False,
         max_function_calls: int = DEFAULT_MAX_FUNCTION_CALLS,
         callback_manager: Optional[CallbackManager] = None,
@@ -55,6 +58,7 @@ class FnRetrieverOpenAIAgent(BaseOpenAIAgent):
         prefix_messages: Optional[List[ChatMessage]] = None,
     ) -> "FnRetrieverOpenAIAgent":
         chat_history = chat_history or []
+        memory = memory or memory_cls.from_defaults(chat_history=chat_history)
         llm = llm or OpenAI(model=DEFAULT_MODEL_NAME)
         if not isinstance(llm, OpenAI):
             raise ValueError("llm must be a OpenAI instance")
@@ -76,7 +80,7 @@ class FnRetrieverOpenAIAgent(BaseOpenAIAgent):
         return cls(
             retriever=retriever,
             llm=llm,
-            chat_history=chat_history,
+            memory=memory,
             prefix_messages=prefix_messages,
             verbose=verbose,
             max_function_calls=max_function_calls,