+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+## Unreleased
+
+### New Features
+- Added basic chat buffer memory to agents / chat engines (#6857)
+
 ## [v0.7.5] - 2023-07-11
 
 ### New Features